+++ Playground.playground/Contents.swift
@@ -35,8 +35,8 @@ class Person : RowModel {
     
     // RowModel overrides
     
-    override class var databaseTable: Table? {
-        return Table(named: "persons", primaryKey: .RowID("id"))
+    override class var databaseTableName: String? {
+        return "persons"
     }
     
     override func setDatabaseValue(dbv: DatabaseValue, forColumn column: String) {
@@ -74,5 +74,15 @@ let persons = dbQueue.inDatabase { db in
     Person.fetchAll(db, "SELECT * FROM persons ORDER BY firstName, lastName")
 }
 
-print(persons)
-print(persons.map { $0.fullName })
+dbQueue.inDatabase { db in
+    for row in Row.fetchAll(db, "PRAGMA index_list('persons')") {
+        print(row)
+    }
+    for row in Row.fetchAll(db, "PRAGMA foreign_key_list('persons')") {
+        print(row)
+    }
+    for row in Row.fetchAll(db, "PRAGMA table_info('persons')") {
+        print(row)
+    }
+    print(Row.fetchAll(db, "PRAGMA table_info('foo')").count)
+}